+++ src/main/java/org/elasticsearch/search/suggest/phrase/StupidBackoffScorer.java
@@ -44,7 +44,7 @@ public class StupidBackoffScorer extends WordScorer {
     @Override
     protected double scoreBigram(Candidate word, Candidate w_1) throws IOException {
         SuggestUtils.join(separator, spare, w_1.term, word.term);
-        final int count = frequency(spare);
+        final long count = frequency(spare);
         if (count < 1) {
             return discount * scoreUnigram(word);
         }
@@ -54,17 +54,17 @@ public class StupidBackoffScorer extends WordScorer {
     @Override
     protected double scoreTrigram(Candidate w, Candidate w_1, Candidate w_2) throws IOException {
         SuggestUtils.join(separator, spare, w_2.term, w_1.term, w.term);
-        final int trigramCount = frequency(spare);
+        final long trigramCount = frequency(spare);
         if (trigramCount < 1) {
             SuggestUtils.join(separator, spare, w_1.term, w.term);
-            final int count = frequency(spare);
+            final long count = frequency(spare);
             if (count < 1) {
                 return discount * scoreUnigram(w);
             }
             return discount * (count / (w_1.frequency + 0.00000000001d));
         }
         SuggestUtils.join(separator, spare, w_1.term, w.term);
-        final int bigramCount = frequency(spare);
+        final long bigramCount = frequency(spare);
         return trigramCount / (bigramCount + 0.00000000001d);
     }
